Better styling system with SASS and new global, colors, fonts SCSS.
Introduced a new styling system where (if preferred), components
can be styled by creating a SCSS file under the package name. e.g.
If we had the components:

    src/error/ErrorMessage.jsx
    src/error/Error404.jsx

Each still have their own class `.ErrorMessage` and `.Error404` (if a
component in their own right) but we create a single file called
`Error.scss`, which is the name of the folder. This file is then placed
in the corresponding style directory where the `error` folder would
appear. In this case, it would be `style/Error.scss` since the `error`
directory is at the root of the source.

I also created the following files:

* `Global.scss` - For style tags, wildcards etc.
* `common/Colors.scss` - For holding the colors of items.
* `common/Fonts.scss` - Describe the fonts for items.
* `common/Spacing.scss` - (Experimental) Global spacing style.