Added Redux, redux-simple-router and redux-devtools.
This was an incredibly convoluted process of mangling crazy API's
made by very smart people. Redux and co. need a good dose of
sensible defaults because I don't want to configure every fucking
thing. Redux-DevTools are used via the cool Chrome extension and
a single line in the `src/index.jsx`.

I'm pretty sure I'm not crazy thinking this is a shite API.