redux-actions, redux-promise and, Login action creator and reducer.
Finally implemented my first action that dispatches when you hit
the login button. I was stuck for a while with an issue where my
action wouldn't dispatch when passed as a prop so I needed to bind
it. It was odd though, because the context shouldn't have had anything
to do with it.

I also added redux-actions, which is a pretty sweet library for
defining action creators and handlers. Along with that, I've added
redux-promise as middleware for some nifty async-ness.