Splitting the application in parts, Store, Main and index.jsx.
To ensure we have the most testability, I've moved the rendering
logic into it's own file, index.jsx and created a Main.jsx which
creates all the routes for the application. I also moved the redux
store creation logic to it's own file, Store.js. To do this, I
had to upgrade react-router to 2.0.0-rc4 which exported a singleton
`browserHistory`. This was then used by both the Main.jsx and Store.js
(redux-simple-router).