Loading module page and login redirect.
A pattern in Redux is starting to appear and this pattern requires
a shitload of work. So much boilerplate everywhere, hopefully it
eases up as we move along. The module page is working with a loading
interface and then displays it when needed. We also redirect to the
correct page after logins which works absolutely beautifully. Dayum.