Paper view completed.
The paper view is now completed which displays the questions and
a little footer that links to the parser to fix the questions. I fixed
a problem which will be good to note for future where when linking in
React router to an already mounted component, it didn't load the content
because the call for loading the content was in the `componentDidMount`.
I had to hook into the componentWillReceiveProps method to load the paper.