Lot's of restructing, i18n skeleton, Signup page and Insititution.
Did lots of restructuring. New rule: Anything that is not connected
to the state is a ui component and should be store in views/ui. Updated
SCSS file structure to more accurately reflect the source files.

Barebones i18n that only supports English so far. It has strings in a
seperate YAML file which loads it as a Javascript object. Simple stuff.

Signup page and Institution design on the way. Using a sort of card style
design which is starting to look quite nice.