Large update. Added new entities and had to move the JSON serialization.
New entity class that uses SQLAlchemy joined-table style inheritance. Each
new "thing" in the database that wants to be commentable or likable can
simply just subclass the Entity and bam, all that functionality is enabled.

The new entites we have are Note, Solution, Comment and now converted, Question.

Note: The JSON serialization logic has to be moved to it's own class due to
the way SQL Alchemy's inheritance uses Subclass. One the root Model (Entity)
would get a schema and any subclass would attempt to use that schema thus
not serializing their own fields. Now, models that wish to be serializable
have to explicitly subclass `Serializable`.